Fix auto_trace_test.cc
Related:
        APPLINK-11127

+++ src/components/utils/test/auto_trace_test.cc
@@ -40,6 +40,7 @@
 #include "utils/file_system.h"
 #include "utils/threads/thread.h"
 #include "utils/date_time.h"
+#include "utils/logger_status.h"
 
 namespace test {
 namespace components {
@@ -50,10 +51,12 @@ using namespace ::logger;
 CREATE_LOGGERPTR_GLOBAL(logger_, "AutoTraceTestLog")
 
 void Preconditions() {
-  // Delete file with previous logs
   const char* file_name = "AutoTraceTestLogFile.log";
-  ASSERT_TRUE(file_system::DeleteFile(file_name))
-    << "Can't delete AutoTraceTestLogFile.log\n";
+  // Delete file with previous logs
+  if (file_system::FileExists(file_name)){
+    ASSERT_TRUE(file_system::DeleteFile(file_name) )
+        << "Can't delete AutoTraceTestLogFile.log\n";
+  }
 }
 
 void InitLogger() {
@@ -112,6 +115,8 @@ TEST(AutoTraceTest, AutoTrace_WriteToFile_ReadCorrectString) {
   DeinitLogger();
 
   ASSERT_TRUE(CheckAutoTraceDebugInFile(testlog));
+  LogMessageLoopThread::destroy();
+  logger::logger_status = LoggerThreadNotCreated;
 }
 
 }  // namespace utils_test
